requirements.txt
absl-py==1.4.0
aio-pika==8.2.3
aioconsole==0.8.1
aiofiles==24.1.0
aiogram==2.15
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiohttp-retry==2.9.0
aiolimiter==1.1.0
aiormq==6.4.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
blinker==1.8.2
boto3==1.35.52
botocore==1.35.52
breadability==0.1.20
bs4==0.0.2
CacheControl==0.12.14
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
class-registry==2.1.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
comm==0.2.2
confluent-kafka==2.6.0
cryptography==43.0.3
cycler==0.12.1
dask==2022.10.2
datasets==3.2.0
debugpy==1.8.11
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
dill==0.3.8
discord.py==2.4.0
distro==1.9.0
dnspython==2.3.0
docopt==0.6.2
emoji==2.14.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
fbmessenger==6.0.0
filelock==3.16.1
filters==1.3.2
fire==0.7.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googletrans==4.0.0rc1
greenlet==3.1.1
grpcio==1.67.1
h11==0.14.0
h2==3.2.0
h5py==3.12.1
hpack==3.0.0
hstspreload==2024.10.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.26.1
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.2.0
json5==0.10.0
jsonpickle==3.0.4
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.6.17
keras==2.12.0
kiwisolver==1.4.7
langdetect==1.0.9
libclang==18.1.1
libretranslatepy==2.1.4
locket==1.0.0
logging==0.4.9.6
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.5.3
matplotlib-inline==0.1.7
mattermostwrapper==2.2
mdurl==0.1.2
mistune==3.0.2
ml_dtypes==0.5.0
msgpack==1.1.0
multidict==5.2.0
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.6.3
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.28.0
opt_einsum==3.4.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pamqp==3.2.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
Pillow==10.1.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.23.3
psutil==6.1.1
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyaes==1.6.1
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==1.10.9
pydantic_core==2.23.4
pydot==1.4.2
Pygments==2.18.0
pyjokes==0.8.3
PyJWT==2.9.0
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.2.0
pyreadline3==3.5.4
pyrsistent==0.20.0
PySocks==1.7.1
python-crfsuite==0.9.11
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.10.1
python-json-logger==3.2.1
python-slugify==8.0.4
python-socketio==5.11.4
python-telegram-bot==21.11.1
pytube==15.0.0
pytz==2022.7.1
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
questionary==1.10.0
randomname==0.1.5
rasa==3.6.20
rasa-sdk==3.6.2
redis==4.6.0
referencing==0.35.1
regex==2022.10.31
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
rocketchat-API==1.30.0
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.12
s3transfer==0.10.3
safetensors==0.4.5
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.10.1
seaborn==0.13.2
selenium==4.29.0
Send2Trash==1.8.3
sentry-sdk==1.14.0
simple-websocket==1.1.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack_sdk==3.33.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.54
stack-data==0.6.3
structlog==23.3.0
structlog-sentry==2.1.0
sumy==0.11.0
tabulate==0.9.0
tarsafe==0.0.4
telegraph==2.2.0
Telethon==1.38.1
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.1
terminaltables==3.1.10
text-unidecode==1.3
textblob==0.18.0.post0
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.20.1
tomli==2.2.1
toolz==1.0.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.2
trio==0.29.0
trio-websocket==0.12.1
twilio==8.2.2
types-python-dateutil==2.9.0.20241206
typing-utils==0.1.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
webexteamssdk==1.6.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.14.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.16.0
